requirements.txt
requests==2.31.0
Flask==2.3.3
psycopg2-binary==2.9.10
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.2